+++ src/ui/framework/widgets/core/box.hpp
@@ -7,10 +7,10 @@ namespace widgets {
 
 struct Rect
 {
-    uint8_t x;
-    uint8_t y;
-    uint8_t w;
-    uint8_t h;
+    uint16_t x;
+    uint16_t y;
+    uint16_t w;
+    uint16_t h;
 };
 
 } // namespace widgets
